Improve public API tests
