+++ e2e/testcases/test-data/birth-declaration.ts
@@ -201,6 +201,8 @@ export async function createDeclaration(
     annotation
   })
 
+  console.log('it does register it??')
+
   const registerAction = registerRes.actions.find(
     (action: ActionDocument) => action.type === 'REGISTER'
   )
